fix(queue): corrigir ROUND em double precision para numeric no findDailyQueue
PostgreSQL não suporta ROUND(double precision, integer), apenas
ROUND(numeric, integer). O erro causava falha no endpoint
/analytics/daily-queue, afetando o componente DailyQueue na DashboardPage.

+++ backend/src/repositories/cardRepository.ts
@@ -101,8 +101,8 @@ class CardRepository {
     const result = await pool.query(
       `SELECT c.id, c.front, c.back, c.stability, c.due, c.state,
               ROUND(
-                EXP(-GREATEST(EXTRACT(EPOCH FROM NOW() - c.due) / 86400, 0)
-                    / COALESCE(NULLIF(c.stability, 0), 1)) * 100, 2
+                (EXP(-GREATEST(EXTRACT(EPOCH FROM NOW() - c.due) / 86400, 0)
+                    / COALESCE(NULLIF(c.stability, 0), 1)) * 100)::numeric, 2
               ) AS predicted_recall
        FROM cards c
        JOIN decks d ON d.id = c.deck_id
